+++ app/models/event.rb
@@ -2,6 +2,12 @@ class Event < ApplicationRecord
   has_many :event_attendees, dependent: :destroy
   has_many :users, through: :event_attendees
 
+  def discontinued?
+    return Time.zone.today > end_date if end_date.present?
+
+    Time.zone.today > start_date
+  end
+
   # :nocov:
   def self.ransackable_attributes(_auth_object = nil)
     %w[address country_code end_date name start_date url]