+++ test/models/product_variant_test.rb
@@ -23,7 +23,7 @@ class ProductVariantTest < ActiveSupport::TestCase
       discontinued_year: 2020,
       discontinued_month: 12
     )
-    assert_equal '2020&thinsp;/&thinsp;12', product_variant.formatted_discontinued_date
+    assert_equal '2020/12', product_variant.formatted_discontinued_date
 
     product_variant.update!(
       discontinued: true,
@@ -31,7 +31,7 @@ class ProductVariantTest < ActiveSupport::TestCase
       discontinued_month: 12,
       discontinued_day: 31
     )
-    assert_equal '2020&thinsp;/&thinsp;12&thinsp;/&thinsp;31', product_variant.formatted_discontinued_date
+    assert_equal '2020/12/31', product_variant.formatted_discontinued_date
   end
 
   test 'name_with_fallback' do